[Issue #10] Ensure words do not shuffle in the original order
TL;DR: If a word gets shuffled into its original order, just shuffle it
manually.

Long Version:
- Implement and test a util function for checking if arrays are equal.
This is used for checking if the shuffled letters are in the original
order.
- Break up the shuffle module into a few smaller functions. The default
export's logic is now a wrapper for the original shuffling behavior.
However, it does the checking of the order of the word shuffling.
- Extract getting a random number generator to an injector. This enables
us to mock what the rng will produce, and thus reproduce the conditions
in which the order is unchanged.
Please enter the commit message for your changes. Lines starting

Miscellaneous:
I also enabled strict null checks, because it occurred to me that this
codebase did not already have it on.